Return matchToKartoffel back into one file
- because of sendLog not defined
